Add auth

+++ backend/SuperCompose/Graphql/Mutation.cs
@@ -10,7 +10,8 @@ using System.Security.Claims;
 using System.Threading;
 using System.Threading.Tasks;
 using HotChocolate;
-using HotChocolate.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Authorization;
+using SuperCompose.Auth;
 using SuperCompose.Context;
 using SuperCompose.Exceptions;
 
@@ -23,13 +24,15 @@ namespace SuperCompose.Graphql
     private readonly DeploymentService deploymentService;
     private readonly IDbContextFactory<SuperComposeContext> ctxFactiry;
     private readonly ConnectionService conn;
+    private readonly IAuthorizationService authorizationService;
 
     public Mutation(
       NodeService nodeService,
       ComposeService composeService,
       DeploymentService deploymentService,
       IDbContextFactory<SuperComposeContext> ctxFactiry,
-      ConnectionService conn
+      ConnectionService conn,
+      IAuthorizationService authorizationService
     )
     {
       this.nodeService = nodeService;
@@ -37,6 +40,7 @@ namespace SuperCompose.Graphql
       this.deploymentService = deploymentService;
       this.ctxFactiry = ctxFactiry;
       this.conn = conn;
+      this.authorizationService = authorizationService;
     }
 
     [UnionType("NodeResult")]
@@ -102,7 +106,7 @@ namespace SuperCompose.Graphql
       }
     }
 
-    [Authorize]
+    [HotChocolate.AspNetCore.Authorization.Authorize]
     public async Task<INodeResult> CreateNode(
       [Required] [MaxLength(255)] string name,
       [Required] [MaxLength(255)] string host,
@@ -149,6 +153,7 @@ namespace SuperCompose.Graphql
       return null;
     }
 
+    [HotChocolate.AspNetCore.Authorization.Authorize]
     public async Task<INodeResult> UpdateNode(
       [Required] Guid id,
       [MaxLength(255)] string name,
@@ -156,11 +161,18 @@ namespace SuperCompose.Graphql
       [MaxLength(255)] string username,
       [Range(1, 65535)] int port,
       string? password,
-      string? privateKey
+      string? privateKey,
+      [GlobalState("ClaimsPrincipal")] ClaimsPrincipal user
     )
     {
       try
       {
+        var auth = await authorizationService.AuthorizeAsync(user, id, new TenantNodeRequirement());
+        if (auth.Failure != null)
+        {
+          throw new UnauthorizedAccessException();
+        }
+        
         await nodeService.Update(
           id,
           name,
@@ -180,15 +192,25 @@ namespace SuperCompose.Graphql
       }
     }
 
-    public async Task<bool> DeleteNode([Required] Guid id)
+    [HotChocolate.AspNetCore.Authorization.Authorize]
+    public async Task<bool> DeleteNode([Required] Guid id,
+      [GlobalState("ClaimsPrincipal")] ClaimsPrincipal user)
     {
+      var auth =
+        await authorizationService.AuthorizeAsync(user, id,
+          new TenantNodeRequirement());
+      if (auth.Failure != null)
+      {
+        throw new UnauthorizedAccessException();
+      }
+      
       await nodeService.Delete(id);
 
       return true;
     }
 
 
-    [Authorize]
+    [HotChocolate.AspNetCore.Authorization.Authorize]
     [UseFirstOrDefault]
     [UseProjection]
     public async Task<IQueryable<Compose>> CreateCompose(
@@ -205,7 +227,7 @@ namespace SuperCompose.Graphql
       return ctxFactiry.CreateDbContext().Composes.Where(x => x.Id == id);
     }
 
-
+    [HotChocolate.AspNetCore.Authorization.Authorize]
     [UseFirstOrDefault]
     [UseProjection]
     public async Task<IQueryable<Compose>> UpdateCompose(
@@ -213,9 +235,18 @@ namespace SuperCompose.Graphql
       [MaxLength(255)] string? name,
       [MaxLength(255)] string? directory,
       bool? serviceEnabled,
-      string? compose
+      string? compose,
+      [GlobalState("ClaimsPrincipal")] ClaimsPrincipal user
     )
     {
+      var auth =
+        await authorizationService.AuthorizeAsync(user, id,
+          new TenantComposeRequirement());
+      if (auth.Failure != null)
+      {
+        throw new UnauthorizedAccessException();
+      }
+      
       await composeService.Update(
         id,
         name,
@@ -227,97 +258,190 @@ namespace SuperCompose.Graphql
       return ctxFactiry.CreateDbContext().Composes.Where(x => x.Id == id);
     }
 
-    public async Task<bool> DeleteCompose([Required] Guid id)
+    [HotChocolate.AspNetCore.Authorization.Authorize]
+    public async Task<bool> DeleteCompose(
+      [Required] Guid id,
+      [GlobalState("ClaimsPrincipal")] ClaimsPrincipal user)
     {
+      var auth =
+        await authorizationService.AuthorizeAsync(user, id,
+          new TenantComposeRequirement());
+      if (auth.Failure != null)
+      {
+        throw new UnauthorizedAccessException();
+      }
+      
       await composeService.Delete(id);
 
       return true;
     }
 
+    [HotChocolate.AspNetCore.Authorization.Authorize]
     [UseFirstOrDefault]
     [UseProjection]
     public async Task<IQueryable<Compose>> RedeployCompose(
-      [Required] Guid id
+      [Required] Guid id,
+      [GlobalState("ClaimsPrincipal")] ClaimsPrincipal user
     )
     {
+      var auth =
+        await authorizationService.AuthorizeAsync(user, id,
+          new TenantComposeRequirement());
+      if (auth.Failure != null)
+      {
+        throw new UnauthorizedAccessException();
+      }
       await composeService.Redeploy(id);
 
       return ctxFactiry.CreateDbContext().Composes.Where(x => x.Id == id);
     }
 
+    [HotChocolate.AspNetCore.Authorization.Authorize]
     [UseFirstOrDefault]
     [UseProjection]
     public async Task<IQueryable<Deployment>> CreateDeployment(
       [Required] Guid node,
-      [Required] Guid compose
+      [Required] Guid compose,
+      [GlobalState("ClaimsPrincipal")] ClaimsPrincipal user
     )
     {
+      var nodeAuth =
+        await authorizationService.AuthorizeAsync(user, node,
+          new TenantNodeRequirement());
+      var deploymentAuth =
+        await authorizationService.AuthorizeAsync(user, node,
+          new TenantDeploymentRequirement());
+      if (nodeAuth.Failure != null || deploymentAuth.Failure != null)
+      {
+        throw new UnauthorizedAccessException();
+      }
+      
       var id = await deploymentService.Create(node, compose);
 
       return ctxFactiry.CreateDbContext().Deployments.Where(x => x.Id == id);
     }
 
+    [HotChocolate.AspNetCore.Authorization.Authorize]
     [UseFirstOrDefault]
     [UseProjection]
     public async Task<IQueryable<Deployment>> EnableDeployment(
-      [Required] Guid deployment
+      [Required] Guid deployment,
+      [GlobalState("ClaimsPrincipal")] ClaimsPrincipal user
     )
     {
+      var deploymentAuth =
+        await authorizationService.AuthorizeAsync(user, deployment,
+          new TenantDeploymentRequirement());
+      if (deploymentAuth.Failure != null)
+      {
+        throw new UnauthorizedAccessException();
+      }
+      
       await deploymentService.Enable(deployment);
 
       return ctxFactiry.CreateDbContext().Deployments.Where(x => x.Id == deployment);
     }
 
+    [HotChocolate.AspNetCore.Authorization.Authorize]
     [UseFirstOrDefault]
     [UseProjection]
     public async Task<IQueryable<Deployment>> DisableDeployment(
-      [Required] Guid deployment
+      [Required] Guid deployment,
+      [GlobalState("ClaimsPrincipal")] ClaimsPrincipal user
     )
     {
+      var deploymentAuth =
+        await authorizationService.AuthorizeAsync(user, deployment,
+          new TenantDeploymentRequirement());
+      if (deploymentAuth.Failure != null)
+      {
+        throw new UnauthorizedAccessException();
+      }
+      
       await deploymentService.Disable(deployment);
 
       return ctxFactiry.CreateDbContext().Deployments.Where(x => x.Id == deployment);
     }
 
+    [HotChocolate.AspNetCore.Authorization.Authorize]
     [UseFirstOrDefault]
     [UseProjection]
     public async Task<IQueryable<Deployment>> RedeployDeployment(
-      [Required] Guid id
+      [Required] Guid id,
+      [GlobalState("ClaimsPrincipal")] ClaimsPrincipal user
     )
     {
+      var deploymentAuth =
+        await authorizationService.AuthorizeAsync(user, id,
+          new TenantDeploymentRequirement());
+      if (deploymentAuth.Failure != null)
+      {
+        throw new UnauthorizedAccessException();
+      }
+      
       await deploymentService.Redeploy(id);
 
       return ctxFactiry.CreateDbContext().Deployments.Where(x => x.Id == id);
     }
 
+    [HotChocolate.AspNetCore.Authorization.Authorize]
     [UseFirstOrDefault]
     [UseProjection]
     public async Task<IQueryable<Node>> EnableNode(
-      [Required] Guid node
+      [Required] Guid node,
+      [GlobalState("ClaimsPrincipal")] ClaimsPrincipal user
     )
     {
+      var deploymentAuth =
+        await authorizationService.AuthorizeAsync(user, node,
+          new TenantNodeRequirement());
+      if (deploymentAuth.Failure != null)
+      {
+        throw new UnauthorizedAccessException();
+      }
+      
       await nodeService.Enable(node);
 
       return ctxFactiry.CreateDbContext().Nodes.Where(x => x.Id == node);
     }
 
+    [HotChocolate.AspNetCore.Authorization.Authorize]
     [UseFirstOrDefault]
     [UseProjection]
     public async Task<IQueryable<Node>> DisableNode(
-      [Required] Guid node
+      [Required] Guid node,
+      [GlobalState("ClaimsPrincipal")] ClaimsPrincipal user
     )
     {
+      var deploymentAuth =
+        await authorizationService.AuthorizeAsync(user, node,
+          new TenantNodeRequirement());
+      if (deploymentAuth.Failure != null)
+      {
+        throw new UnauthorizedAccessException();
+      }
+      
       await nodeService.Disable(node);
 
       return ctxFactiry.CreateDbContext().Nodes.Where(x => x.Id == node);
     }
 
+    [HotChocolate.AspNetCore.Authorization.Authorize]
     [UseFirstOrDefault]
     [UseProjection]
     public async Task<IQueryable<Node>> RedeployNode(
-      [Required] Guid id
+      [Required] Guid id,
+      [GlobalState("ClaimsPrincipal")] ClaimsPrincipal user
     )
     {
+      var deploymentAuth =
+        await authorizationService.AuthorizeAsync(user, id,
+          new TenantNodeRequirement());
+      if (deploymentAuth.Failure != null)
+      {
+        throw new UnauthorizedAccessException();
+      }
+      
       await nodeService.Redeploy(id);
 
       return ctxFactiry.CreateDbContext().Nodes.Where(x => x.Id == id);
